Added timestamp key_val for FOVs, used to store the last time the FOV data has been modified (new ROI, deleted ROI, drift computation). The deletion of a ROI was fixed as the deleted ROI was only removed from the view and not from the database.

+++ src/pydetecdiv/domain/Project.py
@@ -315,9 +315,10 @@ class Project:
         :param dso: the object to delete
         :type dso: object (DomainSpecificObject)
         """
-        if (dso.__class__.__name__, dso.id_) in self.pool:
-            del self.pool[dso.__class__.__name__, dso.id_]
-        self.repository.delete_object(dso.__class__.__name__, dso.id_)
+        if dso is not None:
+            if (dso.__class__.__name__, dso.id_) in self.pool:
+                del self.pool[dso.__class__.__name__, dso.id_]
+            self.repository.delete_object(dso.__class__.__name__, dso.id_)
 
     def get_object(self, class_name: str, id_: int = None, uuid: str = None, use_pool: bool = True) -> DSO:
         """
